+++ gitApp/checks/Authenticator.py
