+++ gitApp/hook/views.py
@@ -1,7 +1,7 @@
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 from hook.helper import pretty_request
-from hook.CheckFlow import CheckFlow
+from checks.CheckFlow import CheckFlow
 import json
 
 # Create your views here.
